updated cmake for windows

+++ build.sh
@@ -1,3 +1,7 @@
+if [[ -d /mingw64/bin ]]
+then
+	export PATH=/mingw64/bin:$PATH
+fi
 rm Dj-Engine
 mkdir build-dir
 cd build-dir
@@ -5,5 +9,7 @@ cmake ..
 make
 cp Dj-Engine ../Dj-Engine
 cd ..
-rm -r build-dir
-DRI_PRIME=1 ./Dj-Engine
+if [ "$1" == "-f" ]
+then
+	rm -r build-dir
+fi
